Parser Behaviour changed. See Description
Change to JSON language definition, commas before group closing
characters ( ] and } ) now do not make the parser break. This is a
change mostly done to allow it to be easier to make JSON files by hand.

Generated JSON will _NEVER_ have these commas, it is simply supported to
load files this way. This means that XtoJSON will load 'not well formed'
JSON.

In the future, this behaviour will be controlled by a compiler flag,
JSON_STRICT_COMMAS.